requirements.txt
alembic==0.9.9
asn1crypto==0.24.0
Babel==2.5.3
backcall==0.1.0
blinker==1.4
cffi==1.11.5
click==6.7
cloudpickle==0.5.2
decorator==4.3.0
dominate==2.3.1
Flask==0.12.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.6.0
flask-shell-ipython==0.3.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.6
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
jupyter-client==5.2.3
jupyter-core==4.4.0
Mako==1.0.7
MarkupSafe==1.0
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.1
PyMySQL==0.8.0
python-dateutil==2.7.2
python-dotenv==0.8.2
python-editor==1.0.3
pytz==2018.4
pyzmq==17.0.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.6
tornado==5.0.2
traitlets==4.3.2
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.14.1
WTForms==2.1
WTForms-SQLAlchemy==0.1
